refactor: move common helper functions to new package
located helper functions to a new common package to be shared with all other packages
