+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "go.sum|package-lock.json|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2023-01-05T15:28:56Z",
+  "generated_at": "2023-01-12T20:53:14Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -82,7 +82,7 @@
         "hashed_secret": "892bd503fb45f6fcafb1c7003d88291fc0b20208",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 104,
+        "line_number": 115,
         "type": "Secret Keyword",
         "verified_result": null
       },
@@ -90,7 +90,7 @@
         "hashed_secret": "d4c3d66fd0c38547a3c7a4c6bdc29c36911bc030",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 165,
+        "line_number": 176,
         "type": "Secret Keyword",
         "verified_result": null
       }
