+++ common/travis_test.go
@@ -1,9 +1,10 @@
-package testhelper
+package common
 
 import (
-	"github.com/stretchr/testify/assert"
 	"strings"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 type travisVariables struct {
