+++ testhelper/general.go
@@ -2,40 +2,19 @@
 package testhelper
 
 import (
-	"encoding/json"
 	"log"
 	"os"
-	"os/exec"
-	"reflect"
-	"strings"
-	"sync"
-	"testing"
-
-	"github.com/stretchr/testify/require"
 
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/cloudinfo"
 )
 
 const ForceTestRegionEnvName = "FORCE_TEST_REGION"
 
-// var lock sync.Mutex // use for thread-safe operations
-
 // TesthelperTerraformOptions options object for optional variables to set
 // primarily used for mocking external services in test cases
 type TesthelperTerraformOptions struct {
-	CloudInfoService              CloudInfoServiceI
+	CloudInfoService              cloudinfo.CloudInfoServiceI
 	ExcludeActivityTrackerRegions bool
-}
-
-// interface for the cloudinfo service (can be mocked in tests)
-type CloudInfoServiceI interface {
-	GetLeastVpcTestRegion() (string, error)
-	GetLeastVpcTestRegionWithoutActivityTracker() (string, error)
-	GetLeastPowerConnectionZone() (string, error)
-	LoadRegionPrefsFromFile(string) error
-	HasRegionData() bool
-	RemoveRegionForTest(string)
-	GetThreadLock() *sync.Mutex
 }
 
 // GetBestVpcRegion is a method that will determine a region available
@@ -167,8 +146,8 @@ func GetBestPowerSystemsRegionO(apiKey string, prefsFilePath string, defaultRegi
 }
 
 // configureCloudInfoService is a private function that will configure and set up a new CloudInfoService for testhelper
-func configureCloudInfoService(apiKey string, prefsFilePath string, options TesthelperTerraformOptions) (CloudInfoServiceI, error) {
-	var cloudSvc CloudInfoServiceI
+func configureCloudInfoService(apiKey string, prefsFilePath string, options TesthelperTerraformOptions) (cloudinfo.CloudInfoServiceI, error) {
+	var cloudSvc cloudinfo.CloudInfoServiceI
 
 	// configure new cloudinfosvc if required (not supplied in options)
 	if options.CloudInfoService != nil {
@@ -204,101 +183,4 @@ func configureCloudInfoService(apiKey string, prefsFilePath string, options Test
 	}
 
 	return cloudSvc, nil
-}
-
-// GetRequiredEnvVars returns a map containing required environment variables and their values
-// Fails the test if any are missing
-func GetRequiredEnvVars(t *testing.T, variableNames []string) map[string]string {
-	var missingVariables []string
-	envVars := make(map[string]string)
-
-	for _, variableName := range variableNames {
-		val, present := os.LookupEnv(variableName)
-		if present {
-			envVars[variableName] = val
-		} else {
-			missingVariables = append(missingVariables, variableName)
-		}
-	}
-	require.Empty(t, missingVariables, "The following environment variables must be set: %v", missingVariables)
-
-	return envVars
-}
-
-// GitRootPath gets the path to the current git repos root directory
-func GitRootPath(fromPath string) (string, error) {
-	cmd := exec.Command("git", "rev-parse", "--show-toplevel")
-	cmd.Dir = fromPath
-	path, err := cmd.Output()
-
-	if err != nil {
-		return "", err
-	}
-	return strings.TrimSpace(string(path)), nil
-}
-
-// GetBeforeAfterDiff takes a JSON string as input and returns a string with the differences
-// between the "before" and "after" objects in the JSON.
-//
-// For example, given the JSON string:
-//
-//	{"before": {"a": 1, "b": 2}, "after": {"a": 2, "b": 3}}
-//
-// the function would return the string:
-//
-//	"Before: {"b": 2}\nAfter: {"a": 2, "b": 3}"
-func GetBeforeAfterDiff(jsonString string) string {
-	// Parse the JSON string into a map
-	var jsonMap map[string]interface{}
-	err := json.Unmarshal([]byte(jsonString), &jsonMap)
-	if err != nil {
-		return "Error: unable to parse JSON string"
-	}
-
-	// Get the "before" and "after" values from the map
-	before, beforeOk := jsonMap["before"]
-	after, afterOk := jsonMap["after"]
-	if !beforeOk || !afterOk {
-		return "Error: missing 'before' or 'after' key in JSON"
-	}
-
-	// Check if the "before" and "after" values are objects
-	beforeObject, beforeOk := before.(map[string]interface{})
-	if !beforeOk {
-		return "Error: 'before' value is not an object"
-	}
-	afterObject, afterOk := after.(map[string]interface{})
-	if !afterOk {
-		return "Error: 'after' value is not an object"
-	}
-
-	// Find the differences between the two objects
-	diffsBefore := make(map[string]interface{})
-	for key, value := range beforeObject {
-		if !reflect.DeepEqual(afterObject[key], value) {
-			diffsBefore[key] = value
-		}
-	}
-
-	// Convert the diffs map to a JSON string
-	diffsJson, err := json.Marshal(diffsBefore)
-	if err != nil {
-		return "Error: unable to convert diffs to JSON"
-	}
-
-	// Find the differences between the two objects
-	diffsAfter := make(map[string]interface{})
-	for key, value := range afterObject {
-		if !reflect.DeepEqual(beforeObject[key], value) {
-			diffsAfter[key] = value
-		}
-	}
-
-	// Convert the diffs map to a JSON string
-	diffsJson2, err := json.Marshal(diffsAfter)
-	if err != nil {
-		return "Error: unable to convert diffs2 to JSON"
-	}
-
-	return "Before: " + string(diffsJson) + "\nAfter: " + string(diffsJson2)
 }
