+++ testhelper/tests.go
@@ -19,6 +19,7 @@ import (
 	"github.com/gruntwork-io/terratest/modules/terraform"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/common"
 )
 
 func skipUpgradeTest(branch string) bool {
@@ -58,7 +59,7 @@ func (options *TestOptions) checkConsistency(plan *terraform.PlanStruct) {
 			changesJson = string(changesBytes)
 		}
 
-		resourceDetails := fmt.Sprintf("Name: %s Address: %s Actions: %s\nDIFF:\n%s\n\nChange Detail:\n%s", resource.Name, resource.Address, resource.Change.Actions, GetBeforeAfterDiff(changesJson), changesJson)
+		resourceDetails := fmt.Sprintf("Name: %s Address: %s Actions: %s\nDIFF:\n%s\n\nChange Detail:\n%s", resource.Name, resource.Address, resource.Change.Actions, common.GetBeforeAfterDiff(changesJson), changesJson)
 
 		var errorMessage string
 		if !options.IgnoreDestroys.IsExemptedResource(resource.Address) {
@@ -101,7 +102,7 @@ func (options *TestOptions) testSetup() {
 	}
 
 	// Ensure always running from git root
-	gitRoot, _ := GitRootPath(".")
+	gitRoot, _ := common.GitRootPath(".")
 
 	// To avoid workspace collisions when running in parallel, ignoring any temp terraform files
 	// NOTE: if it is upgrade test we need hidden .git files
@@ -194,7 +195,7 @@ func (options *TestOptions) RunTestUpgrade() (*terraform.PlanStruct, error) {
 		options.testSetup()
 
 		// from here on we will operate in the temp directory
-		gitRoot, _ := GitRootPath(options.TerraformDir)
+		gitRoot, _ := common.GitRootPath(options.TerraformDir)
 		gitRepo, _ := git.PlainOpen(gitRoot)
 
 		// maintain a reference of current checkout, which might be a detatched PR merge, will be used to switch back later
